Move Terms of Use into pages directory
