+++ react-app/src/pages/TermsOfUse.js
@@ -1,0 +1,79 @@
+export default function TermsOfUse() {
+  return (
+    <div className='site__page'>
+      <div className='site__sub-section terms-of-use'>
+        <h2 className='primary-title'>Terms of Use Agreement (draft)</h2>
+        <h3>General</h3>
+        <p>
+          Information on this website is for general informational and
+          educational purposes. Some sections of this website are intended for
+          particular audiences, such as members of the health care community and
+          the general public. By accessing and using this Web website, you
+          accept, without limitation or qualification, this Terms of Use
+          Agreement.
+        </p>
+        <h3>Mental Health Information</h3>
+        <p>
+          This website may contain information about mental health conditions
+          and treatment. This is not meant as a substitute for advice from
+          qualified mental health or other health care professionals.
+          Information on this website should not be used by patients to
+          self-diagnose a mental health or other condition. Patients should
+          always consult with a mental health or other health care professional
+          for advice or information about a diagnosis and treatment.
+        </p>
+        <p>
+          Many questionnaires on this website are based on patient self-report.
+          As such, all responses should be reviewed and interpreted by a
+          qualified mental health or other health care professional. Although
+          information gathered from tests on this website (factoring in the
+          patient's understanding of the questionaire) can inform treatment and
+          diagnoses, any definitive diagnosis and treatment plan should be based
+          on clinical grounds, including all other relevant biopsychosocial
+          information from the patient.
+        </p>
+        <p>
+          Your use of any questionnaires on this website from third-party
+          resources will be governed by such third party's terms and policies
+          (including its privacy policy) and is at your own risk.
+        </p>
+        <h3>Third-party Materials, Websites, and Links</h3>
+        <p>
+          Merely as a convenience, this website may contain links or references
+          to materials and other websites maintained by third parties over whom
+          eDOT has no control. There may also be thir-party links over whom eDOT
+          has no control that access this website. eDOT makes no warranties,
+          guaranties, or representations of any kind as to the completeness,
+          accuracy, or recentness of any and all information found in such
+          third-party websites, including any third-party mobile app and social
+          media platforms. No recommendation or endorsement by eDOT is implied
+          by inclusion of any third-party links on this website. Links from
+          another webiste to this website do not imply any relationship between
+          eDOT and any third party. Your use of any third-party websites is
+          governed by such third party's terms and policies (including its
+          privacy policy) and is at your own risk.
+        </p>
+        <h3>Indemnification</h3>
+        <p>
+          You agree to indemnify, defend and hold harmless eDOT, its officers,
+          directors, employees, agents, suppliers and third party partners from
+          and against all losses, expenses, damages and costs, including
+          reasonable attorneys' fees, resulting from any violation by you of
+          these Terms of Use.
+        </p>
+        <h3>Governing Laws and Miscellaneous</h3>
+        <p>
+          This Terms of Use Agreement and your use of the website shall be
+          governed by the laws of the United States of America and the State of
+          Utah. Any part, provision, representation or warranty of this
+          Agreement which is prohibited or which is held to be void or
+          unenforceable shall be ineffective to the extent of such prohibition
+          or unenforceability without invalidating the remaining provisions
+          hereof. eDOT reserves the right to alter or delete materials from this
+          Web site at any time at its discretion.
+        </p>
+      </div>
+      <div className='one1rem-ht' />
+    </div>
+  );
+}
